ai-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
py-eureka-client
confluent_kafka
python-dotenv
torch
transformers
whisperx
boto3
accelerate
